refactor: 559 - reduce code duplication
