+++ src/widgets/places/ui/place-item/place-item.tsx
@@ -1,0 +1,19 @@
+import React, { PropsWithChildren } from 'react';
+import classNames from 'classnames/bind';
+
+import styles from './place-item.module.scss';
+
+const cx = classNames.bind(styles);
+
+const PlaceItem = ({ children }: PropsWithChildren) => {
+  return (
+    <span className={cx('place-container')}>
+      <span className={cx('place')}>{children}</span>
+      <span className={cx('divider')} data-testid="divider">
+        *
+      </span>
+    </span>
+  );
+};
+
+export default PlaceItem;
