--- src/widgets/places/ui/places.module.scss
@@ -1,57 +1,0 @@
-.places {
-  white-space: nowrap;
-
-  &.content {
-    overflow: hidden;
-    display: flex;
-
-    padding: 5px 20px;
-
-    border-color: $color-gray-600;
-    border-style: solid;
-    border-width: 1px 0 1px 0;
-
-    .marquee-group {
-      will-change: transform;
-      animation: marquee 30s linear infinite;
-
-      @keyframes marquee {
-        100% {
-          translate: -100% 0;
-        }
-      }
-    }
-  }
-
-  .place-container {
-    overflow: hidden;
-    display: inline-flex;
-    flex-direction: row;
-    align-items: center;
-    justify-content: space-between;
-
-    font-size: 20px;
-    font-weight: 500;
-    line-height: 28px;
-    color: $color-black;
-    letter-spacing: 0;
-
-    @include media-tablet {
-      font-size: 16px;
-      font-weight: 400;
-      line-height: 24px;
-    }
-  }
-
-  .place {
-    display: inline-block;
-    padding: 0 16px;
-  }
-
-  .divider {
-    display: inline-block;
-    height: 12px;
-    font-size: 36px;
-    color: $color-red;
-  }
-}
